Integer Input.java
Requested files: InputOutput1.java (Download)
Maximum number of files: 5
Type of work: Individual work
Write a program to get an integer as input and print it. When the user enters an integer, it is stored in a variable test Integer. NOTE: Bold letters or words are Input and the rest are output statements.

Sample Input:

 56

 Sample Output:

 56