Character Input.java
Requested files: InputOutput3.java (Download)
Maximum number of files: 5
Type of work: Individual work
Write a program to get a character as input and print it. When the user enters a character, it is stored in the variable test Character. NOTE: Bold letters or words are Input and the rest are output statements.

Sample Input:

z

Sample Output:

z