Ascii-I.java
Requested files: InputOutput5.java (Download)
Maximum number of files: 5
Type of work: Individual work
ASCII - American Standard Code for Information Interchange. Write a program to get a character as input and print it's ASCII value. When a character is entered, the character itself  stored.  And when we displayed that value using "int()"  type casting , the entered character is displayed as Ascii value. NOTE: Bold letters or words are Input and rest are output statements.

Sample Input :

g

Sample Output:

103