Perfect Number.java
Program to check whether a given number is a perfect number or not is discussed here. A perfect number is a number which is equal to the sum of its proper positive divisors.

For example, 6 is a perfect number.

The divisors of 6 are 1, 2 and 3.

1 + 2 + 3 = 6.

Input & Output Format:

Input consists of one integer.

Output consists of a string. If it is a perfect number then display "Perfect Number" or else display "Not a Perfect Number".

Sample Input:

6

Sample Output:

Perfect Number